Chore: refactor plot
