+++ frontend/src/Components/information_components/index.js
@@ -6,13 +6,13 @@ import {
     Routes,
     Route
   } from "react-router-dom";
-import {Law_Of_Demand} from "./information_components/law_of_demand";
-import {Law_of_Supply} from "./information_components/law_of_supply";
+import {Law_Of_Demand} from "./law_of_demand";
+import {Law_of_Supply} from "./law_of_supply";
 
 
 
 export function Card_information(props) {
-    const { data_slider, adjust_slider} =  props
+    const { data_slider, adjust_slider, data_Demand} =  props
 
     return (
         <Card id="info">
@@ -32,6 +32,7 @@ export function Card_information(props) {
         <BrowserRouter>
         <Routes>
         <Route path="/law_of_demand" element={<Law_Of_Demand
+                    data_Demand = {data_Demand}
                     data_slider = {data_slider}
                     adjust_slider = {adjust_slider}
         />} />
